requirements.txt
aerich==0.7.2
aiofiles==22.1.0
aiohttp==3.8.6
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
choicelib==0.1.5
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
loguru==0.7.2
multidict==6.0.4
openai==0.28.1
pydantic==1.10.13
PyMySQL==1.1.0
pypika-tortoise==0.1.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.2
tortoise-orm==0.20.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
win32-setctime==1.1.0
yarl==1.9.2
